requirements.txt
Flask
psycopg2-binary
Flask-Session
Flask-WeasyPrint
Werkzeug # For secure password hashing
